Move ProposedNode and builder into own file
Moves ProposedNode into store::proposed_node::{ProposedNode, ProposedNodeBuilder}

Structs are still exposed only in store:: via pub use. This change
also requires getters to be implemented on the structs to be used
throughout the store implementations.
