+++ libsplinter/src/admin/store/builders.rs
@@ -437,67 +437,6 @@ impl CircuitProposalBuilder {
     }
 }
 
-/// Builder for creating a `ProposedNode`
-#[derive(Default, Clone)]
-pub struct ProposedNodeBuilder {
-    node_id: Option<String>,
-    endpoints: Option<Vec<String>>,
-}
-
-impl ProposedNodeBuilder {
-    /// Creates a `ProposedNodeBuider`
-    pub fn new() -> Self {
-        ProposedNodeBuilder::default()
-    }
-
-    /// Returns the unique node ID
-    pub fn node_id(&self) -> Option<String> {
-        self.node_id.clone()
-    }
-
-    /// Returns the list of endpoints for the node
-    pub fn endpoints(&self) -> Option<Vec<String>> {
-        self.endpoints.clone()
-    }
-
-    /// Sets the node ID
-    ///
-    /// # Arguments
-    ///
-    ///  * `node_id` - The unique node ID for node
-    pub fn with_node_id(mut self, node_id: &str) -> ProposedNodeBuilder {
-        self.node_id = Some(node_id.into());
-        self
-    }
-
-    /// Sets the endpoints
-    ///
-    /// # Arguments
-    ///
-    ///  * `endpoints` - The list of endpoints for the node
-    pub fn with_endpoints(mut self, endpoints: &[String]) -> ProposedNodeBuilder {
-        self.endpoints = Some(endpoints.into());
-        self
-    }
-
-    /// Builds the `ProposedNode`
-    ///
-    /// Returns an error if the node ID or endpoints are not set
-    pub fn build(self) -> Result<ProposedNode, BuilderError> {
-        let node_id = self
-            .node_id
-            .ok_or_else(|| BuilderError::MissingField("node_id".to_string()))?;
-
-        let endpoints = self
-            .endpoints
-            .ok_or_else(|| BuilderError::MissingField("endpoints".to_string()))?;
-
-        let node = ProposedNode { node_id, endpoints };
-
-        Ok(node)
-    }
-}
-
 /// Builder for creating a `ProposedService`
 #[derive(Default, Clone)]
 pub struct ProposedServiceBuilder {
